refactor: 입력값 검증을 Application에서 하도록 변경
